requirements.txt
asgiref==3.8.1
Django==3.1.8
djangorestframework==3.13.1
sqlparse==0.5.1
tzdata==2024.2
channels==3.0.1
channels-redis==4.2.0
psycopg2-binary==2.9.10
daphne
djangorestframework-simplejwt
django-redis